Added SEQUENCE_OVERHANG_LEFT and _RIGHT
Implements #14 by adding two new input fields. SEQUENCE_OVERHANG_LEFT is
added to the 5' end of the fwd primer while the reverse complement of
SEQUENCE_OVERHANG_RIGHT is added to the 5' end of the rev primer, such
that SEQUENCE_OVERHANG_LEFT and SEQUENCE_OVERHANG_RIGHT are both
supplied in the same orientation. The overhang sequences are utlized in
calculating SELF_ANY, SELF_ANY_STRUCT, SELF_END, SELF_END_STRUCT,
HAIRPIN, HAIRPIN_STRUCT, COMPL_ANY, COMPL_END, plus the _TH versions of
those outputs, and PRODUCT_SIZE. The TM anG GC_PERCENT calculations have
not been modified, so they will only be based on the 3' portion of the
oligo that binds to the template. Total oligo length remains limited to
36 bp, therefore long overhangs will limit the length of the primer
section of the oligo.